use temp file for hmmdb

+++ lib/Tephra/Command/findtrims.pm
@@ -9,6 +9,7 @@ use Pod::Usage    qw(pod2usage);
 use Capture::Tiny qw(capture_merged);
 use Cwd           qw(abs_path getcwd);
 use File::Copy    qw(copy);
+use File::Path    qw(remove_tree);
 use File::Spec;
 use File::Basename;
 use Tephra -command;
@@ -98,16 +99,23 @@ sub _run_trim_search {
     my ($tephra_hmmdb, $tephra_trnadb) = @{$config}{qw(hmmdb trnadb)};
 
     my $dir = getcwd();
-    my $tmpiname  = 'tephra_transposons_hmmdb_XXXX';
-    my $tmp_hmmdbfh = File::Temp->new( TEMPLATE => $tmpiname,
-				       DIR      => $dir,
-				       SUFFIX   => '.hmm',
-				       UNLINK   => 0);
-    my $tmp_hmmdb = $tmp_hmmdbfh->filename;
-    copy $tephra_hmmdb, $tmp_hmmdb or die "Copy failed: $!";
+    my $tmp_hmmdb;
+    my $using_tephra_db = 0;
+    unless (defined $opt->{hmmdb} && -e $opt->{hmmdb}) { 
+        $using_tephra_db = 1;
+        my $dir = getcwd();
+        my $tmpiname  = 'tephra_transposons_hmmdb_XXXX';
+        my $tmp_hmmdbfh = File::Temp->new( TEMPLATE => $tmpiname,
+                                           DIR      => $dir,
+                                           SUFFIX   => '.hmm',
+                                           UNLINK   => 0);
+        $tmp_hmmdb = $tmp_hmmdbfh->filename;
+        copy $tephra_hmmdb, $tmp_hmmdb or die "Copy failed: $!";
+	#$hmmdb = $tmp_hmmdb;
+    }
 
     my $genome  = $opt->{genome};
-    my $hmmdb   = $opt->{hmmdb} // $tmp_hmmdb;
+    my $hmmdb   = $using_tephra_db ? $tmp_hmmdb : $opt->{hmmdb};
     my $trnadb  = $opt->{trnadb} // $tephra_trnadb;
     my $clean   = $opt->{clean} // 0;
     my $debug   = $opt->{debug} // 0;
@@ -131,9 +139,16 @@ sub _run_trim_search {
     my @suff_args = qq(-db $genome -indexname $index -tis -suf -lcp -ssp -sds -des -dna);
     $trim_search->create_index(\@suff_args, $logfile);
     
-    my $strict_gff  = $trim_search->trim_search({ index => $index, mode => 'strict'  });
-    my $relaxed_gff = $trim_search->trim_search({ index => $index, mode => 'relaxed' });
-    unlink $tmp_hmmdb;
+    #my ($strict_gff, $strict_models) 
+    my $strict_gff
+	= $trim_search->trim_search({ index => $index, mode => 'strict'  });
+    #my ($relaxed_gff, $relaxed_models) 
+     my $relaxed_gff
+	 = $trim_search->trim_search({ index => $index, mode => 'relaxed' });
+    unlink $tmp_hmmdb if $using_tephra_db;
+    #remove_tree($strict_models, { safe => 1 });
+    #remove_tree($relaxed_models, { safe => 1 });
+    unlink $logfile unless -s $logfile;
 
     return ($relaxed_gff, $strict_gff);
 }
